Add loading dialog box
Used by full text render pseudo-plugin, displays a progress bar while rendering
of text layers is taking place on document load.

Note that ideally one would want to load the HTML and self-contained styling
introduced in this commit on-demand and dynamically insert it into the DOM so
as to not pollute the main viewer HTML file.  However, doing so would have
required changes to the webpack configuration that were postponed for now, as
it does not seem to support loading of HTML content at present.